test trained model

+++ main.py
@@ -1,6 +1,7 @@
 import torch
 import torch.nn as nn
 import torch.optim as optim
+import os
 
 from torchvision import transforms, models
 from torch.optim import lr_scheduler
@@ -10,43 +11,11 @@ from datasets.transform.ToTensor import ToTensor
 from datasets.transform.RandomCrop import RandomCrop
 from datasets.transform.Rescale import Rescale
 from torch.utils.data import DataLoader
-
+from PIL import Image
 from utils.visualize import visualize_model
 from model.train import train_model
 
 if __name__ == '__main__':
-    
-    # -------- TEST ---------
-    # Data augmentation and normalization for training
-    # Just normalization for validation
-    # data_transforms = {
-    #     'train': transforms.Compose([
-    #         transforms.RandomResizedCrop(224),
-    #         transforms.RandomHorizontalFlip(),
-    #         transforms.ToTensor(),
-    #         transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
-    #     ]),
-    #     'val': transforms.Compose([
-    #         transforms.Resize(256),
-    #         transforms.CenterCrop(224),
-    #         transforms.ToTensor(),
-    #         transforms.Normalize([0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
-    #     ]),
-    # }
-    #
-    # data_dir = '/home/hydra/Downloads/hymenoptera_data'
-    # image_datasets = {x: datasets.ImageFolder(os.path.join(data_dir, x),
-    #                                           data_transforms[x])
-    #                   for x in ['train', 'val']}
-    # dataloaders = {x: torch.utils.data.DataLoader(image_datasets[x], batch_size=4,
-    #                                               shuffle=True, num_workers=4)
-    #                for x in ['train', 'val']}
-    # dataset_sizes = {x: len(image_datasets[x]) for x in ['train', 'val']}
-    # class_names = image_datasets['train'].classes
-    #
-    # device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    # -------- TEST ---------
-
     device = torch.device('cuda:0' if torch.cuda.is_available() else 'cpu')
 
     pets = PETS(root='data', download=True)
@@ -99,21 +68,43 @@ if __name__ == '__main__':
     model_ft = models.resnet18(pretrained=True)
     num_ftrs = model_ft.fc.in_features
 
+    class_to_idx = pets_datasets['train'].class_to_idx
     num_classes = len(pets_datasets['train'].classes)
     print(f'Number of classes: {num_classes}')
     model_ft.fc = nn.Linear(num_ftrs, num_classes)
 
-    model_ft = model_ft.to(device)
+    trained_model_path = 'data/PETS/trained/state_dict_model.pt'
 
-    criterion = nn.CrossEntropyLoss()
+    if os.path.exists(trained_model_path):
+        model_ft.load_state_dict(torch.load(trained_model_path))
+        model_ft.eval()
 
-    optimizer_ft = optim.SGD(model_ft.parameters(), lr=0.001, momentum=0.9)
+        test_image_path = 'data/PETS/eval/images/IMG_1243.jpg'
+        test_image = Image.open(test_image_path)
 
-    # Decay LR by a factor of 0.1 every 7 epochs
-    exp_lr_scheduler = lr_scheduler.StepLR(optimizer_ft, step_size=7, gamma=0.1)
+        with torch.no_grad():
+            outputs = model_ft(transforms.ToTensor()(test_image).unsqueeze(0))
+            _, preds = torch.max(outputs, 1)
 
-    model_conv = train_model(model_ft, dataloaders, dataset_sizes, criterion, optimizer_ft, exp_lr_scheduler, device, num_epochs=25)
+            print('-' * 10 + ' Result ' + '-' * 10)
+            result_dict = {list(class_to_idx.keys())[i]: output.item() for i, output in enumerate(outputs[0])}
 
-    torch.save(model_conv.state_dict(), 'state_dict_model.pt')
+            dict(sorted(result_dict.items(), key=lambda item: item[1]))
+            for item in result_dict:
+                print(f'{item}: {result_dict[item]}')
 
-    visualize_model(model_conv, dataloaders, list(pets_datasets['val'].classes), device, 6)
+    else:
+        criterion = nn.CrossEntropyLoss()
+
+        optimizer_ft = optim.SGD(model_ft.parameters(), lr=0.001, momentum=0.9)
+
+        # Decay LR by a factor of 0.1 every 7 epochs
+        exp_lr_scheduler = lr_scheduler.StepLR(optimizer_ft, step_size=7, gamma=0.1)
+
+        model_conv = train_model(model_ft, dataloaders, dataset_sizes, criterion, optimizer_ft, exp_lr_scheduler,
+                                 device, num_epochs=25)
+
+        torch.save(model_conv.state_dict(), 'state_dict_model.pt')
+
+        visualize_model(model_conv, dataloaders, list(pets_datasets['val'].classes), device, 6)
+
